--- TeamCode/src/main/java/unittests/auto/framework/InternalCameraTest.java
@@ -1,46 +1,0 @@
-package unittests.auto.framework;
-
-import autoutil.vision.TeamElementScanner;
-import unittests.auto.AutoUnitTest;
-
-import static global.General.bot;
-import static global.General.log;
-
-public class InternalCameraTest extends AutoUnitTest {
-    /**
-     * Scanner for team element
-     */
-    TeamElementScanner teamElementScanner = new TeamElementScanner();
-
-    /**
-     * Team element scanner setting and starting
-     */
-    @Override
-    public void init() {
-
-        camera.setInternalScanner(teamElementScanner);
-
-        camera.startInternalCamera();
-    }
-
-    /**
-     * Loop through and log fps
-     */
-    @Override
-    protected void run() {
-        whileActive(() -> {
-            log.show("Internal Camera FPS",
-                    camera.getInternalFPS());
-            log.show("Detected Case", teamElementScanner.getCase());
-        });
-    }
-
-    /**
-     * Stop internal camera
-     */
-    @Override
-    public void stop() {
-
-        camera.stopInternalCamera();
-    }
-}
