--- TeamCode/src/main/java/unittests/tele/framework/IteratorTest.java
@@ -1,4 +1,0 @@
-package unittests.tele.framework;
-
-public class IteratorTest {
-}
