--- TeamCode/src/main/java/unittests/tele/framework/motion/OdometryTest.java
@@ -1,4 +1,0 @@
-package unittests.tele.framework.motion;
-
-public class OdometryTest {
-}
